requirements.txt
PyOpenSSL
python_jwt
requests